Update to Slug Population
When generating any new products the Slug will auto populate with the title of the product. For example, "Party Hat" will have a slug of "party-hat". It looks nice in the URL and is friendly to users. However, I suspect I'll need to remove it in the future because auto incrementing with a number is easier to work with.